requirements.txt
youtube_transcript_api
streamlit
google-generativeai
fpdf
io
re
os
tempfile
python-dotenv
